Move CompilerSearchVisitor into a shared utility in MetalsGlobal
